LayoutTests/http/tests/inspector-enabled/console-clear-arguments-on-frame-navigation.html

<html>
<head>
<script src="resources/console-clear-arguments-test.js"></script>
<script>

if (window.layoutTestController) {
    layoutTestController.waitUntilDone();
    layoutTestController.dumpAsText();
}

var loadCount = 0;
function frameLoaded(event) {
    ++loadCount;
    if (loadCount === 1) {
        document.getElementById('theFrame').src = "about:blank";
        return;
    }

    dumpConsoleMessageArgumentCounts();
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

</script>

</head>
<body>
<p>
Tests that Web Inspector will discard console message arguments when iframe where the message was
logged is navigated to a different page.
</p>
<iframe id="theFrame" onload="frameLoaded(event)" src="resources/console-clear-arguments-iframe.html"></iframe>
<div id="output"></div>
</body>
</html>
